+++ CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Access controlled mintable & burnable LinkToken, for use on sidechains and L2 networks.
+- Added versioning to v0.6 & v0.7 contracts
 
 ### Changed
 
 - Migrated to Hardhat and Waffle
+- Changed LinkToken v0.6 name from 'ChainLink Token' to 'Chainlink Token'
 
 ### Fixed